refactor: move matching logic
